Allow orchestration artifacts to be cleaned up (#253)
Why
---
The `bin/reset` script was included with Hubs Compose instead of
`bin/clean` because it wasn’t obvious whether or not such a script
should remove cloned source code repositories.  The orchestration in
Dash does not clone repositories, so it is not bound by the same
limitations.  I mistakenly included `bin/reset` in the orchestration
scripts when I prepared Dash for orchestration.  This change remedies
that mistake.

With this change an orchestration reset can be performed by running
`bin/clean` and `bin/init` in sequence.

What
----
* Replace `bin/reset` with `bin/clean`
* Update README