Fix Docker build
Why
---
After adding Sentry, the build step wouldn’t complete.  This is because
the sentry dependency compilation scans the entire disk when the project
directory is `/`.  In a docker container the scan hits `/proc` and
recursively follows the `/proc/self/` chain _ad infinitum_.

https://github.com/getsentry/sentry-elixir/issues/183
https://github.com/getsentry/sentry-elixir/issues/168

What
----
Change the project directory in the build stage from `/` to `/src`