Unblock FxA broker on email change (#411)
Why
---
There is a validation on email change that an account with a hub has an
email address stored.  However, some accounts with hubs exist without a
stored email address.  If the validation is necessary it should be
enforced when the record is persisted, not just accessed.  This failing
validation is blocking the FxA broker, which continues to retry the
event.

What
----
* Since email updates in Reticulum are broken anyway, remove the
  aforementioned validation